python-agent/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
asyncio>=3.4.3
typing-extensions>=4.8.0
openai>=1.10.0,<2.0.0
azure-identity>=1.15.0
pyjwt>=2.8.0
aiohttp>=3.9.0
